+++ src/includes/classes/Core/Base/Facades.php
@@ -58,6 +58,7 @@ abstract class Facades
     use Traits\Facades\RequestType;
     use Traits\Facades\Serializer;
     use Traits\Facades\Sha1Mods;
+    use Traits\Facades\SimpleExpressions;
     use Traits\Facades\Slashes;
     use Traits\Facades\Slugs;
     use Traits\Facades\Spellcheck;