+++ tests/local/urls.php
@@ -8,4 +8,5 @@ require_once dirname(__FILE__, 2).'/includes/local.php';
 
 /* ------------------------------------------------------------------------------------------------------------------ */
 
-echo c::appUrl('/client-s/test.js?v=34#core');
+echo c::appUrl('/client-s/test.js?v=34#core')."\n";
+echo c::addUrlQueryArgs(['a' => '', 'b' => '0', 'c' => null], 'https://example.com/');